backend/requirements.txt
fastapi
uvicorn[standard]
lancedb
google-cloud-firestore
google-cloud-aiplatform
firebase-admin
pandas
pydantic
pydantic-settings
python-multipart
python-dotenv
pytest
pytest-asyncio
httpx
google-generativeai
pyarrow
